Refactor auth module
