+++ application/packages/nest-auth/src/index.ts
@@ -5,7 +5,7 @@ export * from './decorators/auth/auth.decorator'
 export * from './guards/auth/auth.guard'
 export * from './guards/access-token/access-token.guard'
 export * from './interfaces/active-user.interface'
-export * from './interfaces/auth-module-options.interface'
+export * from './library/auth-module-options'
 export * from './middleware/token/token.middleware'
 export * from './services/token/token.service'
 export * from './services/token/token.service.mock'
