+++ application/packages/nest-auth/src/library/auth.global-providers.ts
