+++ application/services/passport/src/app/app.module.ts
@@ -3,7 +3,7 @@ import {MicroserviceModule} from '@looport/nest-microservice'
 import {Module} from '@nestjs/common'
 import {ConfigModule as NestConfigModule} from '@nestjs/config'
 
-import {GLOBAL_PROVIDERS} from '@/app/common/app.global-providers'
+import {GLOBAL_PROVIDERS} from '@/app/library/app.global-providers'
 import {AuthenticationModule} from '@/authentication/authentication.module'
 import {ConfigModule} from '@/config/config.module'
 import {getAuthModuleAsyncOptions} from '@/config/services/auth-config/auth-module-options'
