+++ application/services/passport/src/config/services/auth-config/auth-module-options.ts
@@ -1,9 +1,9 @@
-import {AuthModuleAsyncOptions} from '@looport/nest-auth'
+import {ASYNC_OPTIONS_TYPE} from '@looport/nest-auth'
 
 import {ConfigModule} from '@/config/config.module'
 import {AuthConfigService} from '@/config/services/auth-config/auth-config.service'
 
-export const getAuthModuleAsyncOptions = (): AuthModuleAsyncOptions => ({
+export const getAuthModuleAsyncOptions = (): typeof ASYNC_OPTIONS_TYPE => ({
   imports: [ConfigModule],
   inject: [AuthConfigService],
   useFactory: (authenticationConfigService: AuthConfigService) =>
