+++ application/services/telegraph/src/app/app.module.ts
@@ -3,10 +3,10 @@ import {MicroserviceModule} from '@looport/nest-microservice'
 import {Module} from '@nestjs/common'
 import {ConfigModule as NestConfigModule} from '@nestjs/config'
 
-import {GLOBAL_PROVIDERS} from '@/app/common/app.global-providers'
+import {GLOBAL_PROVIDERS} from '@/app/library/app.global-providers'
 import {ConfigModule} from '@/config/config.module'
 import {getAuthModuleAsyncOptions} from '@/config/services/auth-config/auth-module-options'
-import {getMicroserviceModuleAsyncOptions} from '@/config/services/microservcies-config/microservice-module-options'
+import {getMicroserviceModuleAsyncOptions} from '@/config/services/microservcie-config/microservice-module-options'
 import {RoomModule} from '@/room/room.module'
 import {UserModule} from '@/user/user.module'
