+++ application/services/telegraph/src/config/services/microservcie-config/microservice-module-options.ts
@@ -1,7 +1,7 @@
 import {ASYNC_OPTIONS_TYPE} from '@looport/nest-microservice'
 
 import {ConfigModule} from '@/config/config.module'
-import {MicroserviceConfigService} from '@/config/services/microservcies-config/microservice-config.service'
+import {MicroserviceConfigService} from '@/config/services/microservcie-config/microservice-config.service'
 
 export const getMicroserviceModuleAsyncOptions =
   (): typeof ASYNC_OPTIONS_TYPE => ({
